revert reorg of tests for banned deps script
